+++ src/main/java/com/example/booking/services/HabitacionService.java
@@ -18,10 +18,7 @@ import java.nio.file.Files;
 import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Date;
-import java.util.List;
-import java.util.UUID;
-import java.util.Set;
+import java.util.*;
 
 @Service
 public class HabitacionService {
@@ -122,7 +119,7 @@ public class HabitacionService {
 
     public List<Habitacion> habitacionfiltro(Integer id_hotel, Date fecha_inicio, Date fecha_fin, Integer capacidad, boolean wifi1,
                                         boolean terraza1, boolean tv1, boolean aire1, boolean banio_privado, boolean cocina1,
-                                        boolean caja_fuerte, Integer precioBase, Integer puntuacion) {
+                                        boolean caja_fuerte, Double precioBase, Integer puntuacion) {
         List<Integer> wifi = new ArrayList<>();
         List<Integer> terraza = new ArrayList<>();
         List<Integer> tv = new ArrayList<>();
@@ -152,7 +149,7 @@ public class HabitacionService {
             estrellamax = 5.1;
         }
         if(precioBase == null){
-            precioBase = 5;
+            precioBase = 5.0;
             precioFinal = 10000;
 
         }
