+++ src/main/java/com/example/booking/services/HotelService.java
@@ -19,6 +19,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
@@ -37,7 +38,7 @@ public class HotelService {
 
     public List<Hotel> buscarPorFiltros(String ciudad, Date fecha_inicio, Date fecha_fin, Integer capacidad, boolean wifi1,
                                         boolean terraza1, boolean tv1, boolean aire1, boolean banio_privado, boolean cocina1,
-                                        boolean caja_fuerte, Integer precioBase, Integer puntuacion) {
+                                        boolean caja_fuerte, Double precioBase, Integer puntuacion) {
         List<Integer> wifi = new ArrayList<>();
         List<Integer> terraza = new ArrayList<>();
         List<Integer> tv = new ArrayList<>();
@@ -67,7 +68,7 @@ public class HotelService {
             estrellamax = 5.1;
         }
         if(precioBase == null){
-            precioBase = 5;
+            precioBase = 5.0;
             precioFinal = 10000;
 
         }
